Add `read` and use in compile-file.
This adds an implementation of `read` which is used as a replacement
for Racket's `read` in the front-end to parse s-expressions, which is
important for bootstrapping since we will eventually need to be able
to compile the reader.

The implementation uses `read-char` and `peek-char` and is fairly
straightforward recursive descent parser written as a set of mutually
recursive functions.