Draft of support for variable arity functions (#29)
* first attempt

* more work done, bus error sad face

* fixing up some stuff, but still segfaulting, will try to merge with main now and see if it fixes anything

* Fix syntax error in interp.

* fix things up, ready now question mark

* Callee pop args

Previously the calleer of a function pushes arguments on the stack
then pops them off.  This changes the callee to do the popping.  This
should be helpful for variable arity functions and tails; in both
cases, the callee needs to pop arguments off and the caller won't know
how many.

* Fix up variable arity function definition.

Now that callees pop arguments, variable arity is pretty simple.  If
the function takes n mandatory arguments and is given m arguments,
check that m >= n and pop m-n args off, constructing a list, then push
the list back on, so the list has n+1 args on the stack.  Before
returning, pop n+1 args off.  Only remaining wrinkle is that the
return address sits below the args so you have to pop and restore it
before doing these manipulations.

Co-authored-by: David Van Horn <dvanhorn@cs.umd.edu>